% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coefficient_variables.R
\name{create_coeffs_variables_drm_BC5}
\alias{create_coeffs_variables_drm_BC5}
\title{Create Coefficient Variables}
\usage{
create_coeffs_variables_drm_BC5(x_new, coeff, index)
}
\arguments{
\item{x_new}{The sequence list of new x values.}

\item{coeff}{The collection of coefficient values produced by the model.}

\item{index}{The column index used to identify the curve variable}
}
\value{
A collection of coefficient variables
}
\description{
Create Coefficient Variables
}
