% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_sequence.R
\name{create_new_sequence}
\alias{create_new_sequence}
\title{Create New Sequenced List}
\usage{
create_new_sequence(min_x, max_x, increment = 10)
}
\arguments{
\item{min_x}{The min value of 'x'}

\item{max_x}{The max value of 'x'}

\item{increament}{The increment used to create the sequenced list.}
}
\value{
A sequenced list starting at min_x and increasing by increment to
max_x
}
\description{
Creates a list that increased from the minimum 'x' value by the increment
value until max_x is reached.
}
