% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_fit_to_model.R
\name{is_data_fit_to_model}
\alias{is_data_fit_to_model}
\title{Is Data Fit to Model}
\usage{
is_data_fit_to_model(df, has_adstock_data)
}
\arguments{
\item{df}{A dataframe consisting of at least a decompMedia column and a
SaturatedSpend column that will be used to create an scurve.}

\item{has_adstock_data}{Boolean value to indicate if the dataframe consists
of a column called AdstockMedia}
}
\value{
A boolean value indicating if the data is fit to model.
}
\description{
Determines if the model is fit to run through the model based on length of
the data and the max and min values.
}
