% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_drm_curve.R
\name{parse_drm_curve}
\alias{parse_drm_curve}
\title{Parse Coefficients from Curve}
\usage{
parse_drm_curve(curve)
}
\arguments{
\item{curve}{The results from the Modeled data.}
}
\value{
A list of all coefficients for a drm curve
}
\description{
Using the result of the modeled data, this method parses out the coefficients
that are needed throughout the creation of the s-curve.
}
