% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_value.R
\name{replicate_value}
\alias{replicate_value}
\title{Replicate values}
\usage{
replicate_value(value, n)
}
\arguments{
\item{value}{The field that will be replicated}

\item{x_new}{The amount of replications that will be performed}
}
\value{
List of length 'n' consisting entirely of 'value'
}
\description{
Method is used to create a list of 'n' length consisting of 'value'
}
