% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scurve_model.R
\name{run_scurve_model_drm_BC5}
\alias{run_scurve_model_drm_BC5}
\title{Run S-curve Model BC5}
\usage{
run_scurve_model_drm_BC5(data, index)
}
\arguments{
\item{data}{\itemize{
\item the dataframe containing the spend and (DecompMedia or AdstockMedia).
}}

\item{index}{\itemize{
\item column index used to identify the curve variable
}}
}
\value{
The results of the model. Contains the coefficients, summary, and output values.
}
\description{
Using the aomisc library function drm and the BC.5 model approach,
this function will return an s-curve that is fitted to the data provided.
}
