% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid_drm_BC}
\alias{sigmoid_drm_BC}
\title{Sigmoid for the DRM BC model}
\usage{
sigmoid_drm_BC(b, c, d, e, f, x_new)
}
\arguments{
\item{b}{value of the slope at the inflection point}

\item{c}{lower asymptote value}

\item{d}{higher asymptote}

\item{e}{xmid}

\item{x_new}{can be a list or a single value used to determine 'y'}
}
\value{
A 'y' value or list of 'y' values generated by the sigmoid function
}
\description{
The sigmoid function is used to determine a 'y' value for any given 'x'.
}
