% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid_drm_BC5}
\alias{sigmoid_drm_BC5}
\title{Sigmoid for the DRM BC model}
\usage{
sigmoid_drm_BC5(coeff, x_new)
}
\arguments{
\item{coeff}{collection of coefficients}

\item{x_new}{can be a list or a single value used to determine 'y'}
}
\value{
A 'y' value or list of 'y' values generated by the sigmoid function
}
\description{
The sigmoid function is used to determine a 'y' value for any given 'x'.
}
